refactor: move 13 tools_*.py modules into server/tools/ subpackage
Update server/__init__.py bulk import block and _lifespan.py deferred
import to use the new autoskillit.server.tools.* paths.
