+++ src/autoskillit/server/__init__.py
@@ -64,6 +64,8 @@ from autoskillit.core import PIPELINE_FORBIDDEN_TOOLS  # noqa: E402, F401
 from autoskillit.server import (  # noqa: E402, F401
     _misc,
     _notify,
+)
+from autoskillit.server.tools import (  # noqa: E402, F401
     tools_ci,
     tools_ci_merge_queue,
     tools_ci_watch,
@@ -80,7 +82,7 @@ from autoskillit.server import (  # noqa: E402, F401
 )
 from autoskillit.server._factory import make_context  # noqa: E402, F401
 from autoskillit.server._session_type import _apply_session_type_visibility  # noqa: E402, F401
-from autoskillit.server.tools_kitchen import _build_tool_category_listing  # noqa: E402, F401
+from autoskillit.server.tools.tools_kitchen import _build_tool_category_listing  # noqa: E402, F401
 
 # Apply global visibility transform: all sessions start with kitchen tools hidden.
 # Must appear after all tool module imports so the registered tools are in place.
