+++ src/autoskillit/server/tools/__init__.py
@@ -1,0 +1,1 @@
+"""Server tool modules — @mcp.tool() handlers registered on import."""
